// src/api/codegen/schemas/informedTemplateNodeListItem.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ShowDoc 导出接口 OpenAPI
 * 根据 ShowDoc 原始页面自动生成的 OpenAPI 3.0 规范。
 * OpenAPI spec version: 1.0.0
 */
import type { InformedTemplateNodeListItemElementsItem } from './informedTemplateNodeListItemElementsItem'

export interface InformedTemplateNodeListItem {
  node_id: string
  node_name: string
  code: string
  remark: string
  has_parent_node: boolean
  has_child_node: boolean
  child_nodes: InformedTemplateNodeListItem[]
  has_element: boolean
  elements: InformedTemplateNodeListItemElementsItem[]
  required: boolean
  min_limit_str: string
  is_array: boolean
  max_limit_str: string
}