// src/api/codegen/schemas/informedTemplateParagraphListItem.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * ShowDoc 导出接口 OpenAPI
 * 根据 ShowDoc 原始页面自动生成的 OpenAPI 3.0 规范。
 * OpenAPI spec version: 1.0.0
 */
import type { InformedTemplateNodeListItem } from './informedTemplateNodeListItem'

export interface InformedTemplateParagraphListItem {
  paragraph_id: string
  medical_id: string
  code: string
  name: string
  sort: number
  remark: string
  child_paragraph_list: InformedTemplateParagraphListItem[]
  node_list: InformedTemplateNodeListItem[]
}